requirements.txt
mujoco
pytorch_lightning
numpy-stl
vtk
patchelf
termcolor
torchgeometry
scikit-image
numpy
scipy
ipdb
joblib>=1.2.0
opencv-python==4.6.0.66
tqdm 
pyyaml
wandb
scikit-image
gym
git+https://github.com/ZhengyiLuo/smplx.git@master
lxml
human_body_prior
autograd
scikit-learn
chumpy
patchelf
rl-games==1.1.4
wandb
pyvirtualdisplay
chumpy
patchelf
lxml
ipdb
chardet
cchardet
imageio-ffmpeg
easydict
gdown